Refactor pooling2d gtests: extract common code to pooling2d_common.hpp
- Created pooling2d_common.hpp with shared structs, functions, and base class
- Updated pooling2d.cpp, pooling2d_asymmetric.cpp, pooling2d_wide.cpp to use common header
- Moved pooling2d.hpp (ctest driver) back to test/ directory for reference
- All three test files now inherit from Pooling2dCommon<T> base class
- Removed ctest driver dependencies from gtest files
