+++ projects/miopen/test/pooling2d.hpp
@@ -1,7 +1,7 @@
 // Copyright © Advanced Micro Devices, Inc., or its affiliates.
 // SPDX-License-Identifier:  MIT
 
-#include "../pooling_common.hpp"
+#include "pooling_common.hpp"
 
 #define WORKAROUND_ISSUE_1670 1
 #define TEST_GET_INPUT_TENSOR 0
